Attempt to separate peer from state
