+++ proxy/api/src/http/source.rs
@@ -145,7 +145,7 @@ mod handler {
         let default_branch = match peer_id {
             Some(peer_id) if peer_id != ctx.peer_api.peer_id() => {
                 git::Branch::remote(project.default_branch(), &peer_id.to_string())
-            }
+            },
             Some(_) | None => git::Branch::local(project.default_branch()),
         };
 
@@ -281,7 +281,7 @@ mod handler {
         let default_branch = match peer_id {
             Some(peer_id) if peer_id != ctx.peer_api.peer_id() => {
                 git::Branch::remote(project.default_branch(), &peer_id.to_string())
-            }
+            },
             Some(_) | None => git::Branch::local(project.default_branch()),
         };
