+++ proxy/api/src/session.rs
@@ -51,11 +51,11 @@ pub async fn settings(store: &kv::Store) -> Result<settings::Settings, error::Er
 ///
 /// Errors if access to the session state fails, or associated data like the [`identity::Identity`]
 /// can't be found.
-pub async fn current(api: &coco::Api, store: &kv::Store) -> Result<Session, error::Error> {
+pub async fn current(state: &coco::State, store: &kv::Store) -> Result<Session, error::Error> {
     let mut session = get(store, KEY_CURRENT)?;
 
     if let Some(id) = session.identity.clone() {
-        identity::get(api, &id.urn)?;
+        identity::get(state, &id.urn)?;
         session.identity = Some(id);
     }
