+++ proxy/coco/src/project/create.rs
@@ -12,7 +12,7 @@ use radicle_surf::vcs::git::git2;
 
 use crate::config;
 use crate::error::Error;
-use crate::peer;
+use crate::user::User;
 
 /// The data required to either open an existing repository or create a new one.
 #[derive(Debug, Clone, Serialize, Deserialize)]
@@ -165,7 +165,7 @@ impl<Path: AsRef<path::Path>> Create<Path> {
         Ok(repo)
     }
 
-    /// Build a [`project::Project`], where the provided [`peer::User`] is the owner, and the set of
+    /// Build a [`project::Project`], where the provided [`User`] is the owner, and the set of
     /// keys starts with the provided [`keys::PublicKey`].
     ///
     /// # Errors
@@ -173,7 +173,7 @@ impl<Path: AsRef<path::Path>> Create<Path> {
     ///   * Failed to build the project entity.
     pub fn build(
         &self,
-        owner: &peer::User,
+        owner: &User,
         key: keys::PublicKey,
     ) -> Result<project::Project<entity::Draft>, Error> {
         let name = self.repo.project_name()?;
